convert classes to functions
