+++ jwst/outlier_detection/tests/test_algorithms.py
@@ -4,8 +4,8 @@ import pytest
 import numpy as np
 import scipy.signal
 
-from jwst.outlier_detection.outlier_detection_ifu import medfilt
-from jwst.outlier_detection.outlier_detection_tso import moving_median_over_zeroth_axis
+from jwst.outlier_detection.ifu import medfilt
+from jwst.outlier_detection.tso import moving_median_over_zeroth_axis
 
 
 @pytest.mark.parametrize("shape,kern_size", [
